changed one massive file into 3 medium sized files
